Moved AWS storage integration and provider into their own files
