+++ test/storageintegration/storage_integration_provider_tests_create.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import Mock, call
 
 from pulumi_snowflake.storageintegration.storage_integration import StorageIntegration
-from pulumi_snowflake.storageintegration.storage_integration_provider import AWSStorageIntegrationProvider
+from pulumi_snowflake.storageintegration.aws_storage_integration_provider import AWSStorageIntegrationProvider
 
 
 class StorageIntegrationProviderTests(unittest.TestCase):
